+++ TrackApplication/src/test/java/org/edwinsoto/trackapplication/dao/InMemoryArtistDAOTest.java
@@ -1,11 +1,7 @@
-package org.edwinsoto.trackapplication.repository;
+package org.edwinsoto.trackapplication.dao;
 
 import org.edwinsoto.trackapplication.model.Artist;
-import org.edwinsoto.trackapplication.model.Track;
-import org.edwinsoto.trackapplication.service.ArtistService;
 import org.edwinsoto.trackapplication.service.TrackService;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
@@ -14,9 +10,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
 
 import java.time.LocalDate;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 import static org.junit.jupiter.api.Assertions.*;
