+++ TrackApplication/src/test/java/org/edwinsoto/trackapplication/dao/NetworkPricingDAOTest.java
@@ -1,4 +1,4 @@
-package org.edwinsoto.trackapplication.repository;
+package org.edwinsoto.trackapplication.dao;
 
 
 import org.edwinsoto.trackapplication.model.Track;
@@ -7,12 +7,9 @@ import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 
-import org.mockito.InjectMocks;
 import org.springframework.boot.test.mock.mockito.MockBean;
 
 import org.springframework.test.context.junit.jupiter.SpringExtension;
-
-import static org.mockito.Mockito.when;
 
 
 @ExtendWith(SpringExtension.class)
